requirements.txt
Flask==1.0.1
Flask-SQLAlchemy==2.3.2
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.4
SQLAlchemy==1.2.13
Werkzeug==0.14.1
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
idna==2.7
itsdangerous==1.1.0
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pyOpenSSL==18.0.0
pycparser==2.19
pytest==3.9.3
requests==2.19.1
setuptools==39.0.1
six==1.11.0
urllib3==1.23
virtualenv==16.1.0
virtualenv-clone==0.4.0